+++ b1ack_memory/__init__.py
@@ -2,5 +2,5 @@
 
 from .service import MemoryService
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __all__ = ["MemoryService", "__version__"]